+++ src/main/java/com/kl/grooveo/boundedContext/member/controller/restMemberController.java
@@ -1,27 +1,29 @@
 package com.kl.grooveo.boundedContext.member.controller;
 
+import com.kl.grooveo.base.email.service.EmailService;
 import com.kl.grooveo.base.rq.Rq;
 import com.kl.grooveo.boundedContext.follow.service.FollowService;
 import com.kl.grooveo.boundedContext.member.dto.MemberProfileDto;
 import com.kl.grooveo.boundedContext.member.entity.Member;
 import com.kl.grooveo.boundedContext.member.service.MemberService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/api")
 public class restMemberController {
 
     private final MemberService memberService;
     private final FollowService followService;
+    private final EmailService emailService;
     private final Rq rq;
 
-    @GetMapping("api/member/getProfile")
+    @GetMapping("/member/getProfile")
     @ResponseBody
     public MemberProfileDto getProfile(@RequestParam String username) throws Exception {
         Optional<Member> opFollowingUser = memberService.findByUsername(username);
@@ -45,5 +47,23 @@ public class restMemberController {
                 .email(followingUser.getEmail())
                 .isFollowing(isFollowing)
                 .build();
+    }
+
+    @PostMapping("/sendCode")
+    @ResponseBody
+    public String sendVerificationCode(HttpServletRequest request, String userEmail) {
+        HttpSession session = request.getSession();
+
+        emailService.sendVerificationCode(session, userEmail);
+
+        return "";
+    }
+
+    @PostMapping("/certification")
+    @ResponseBody
+    public boolean checkVerificationCode(HttpServletRequest request, String userEmail, String inputCode) {
+        HttpSession session = request.getSession();
+
+        return emailService.emailCertification(session, userEmail, inputCode);
     }
 }
