+++ src/main/resources/templates/usr/member/join.html
@@ -202,7 +202,7 @@
 
             $.ajax({
                 type: "POST",
-                url: "/usr/member/sendCode",
+                url: "/api/sendCode",
                 data: {"userEmail": userEmail},
                 beforeSend: function (xhr) {
                     xhr.setRequestHeader(header, token);
@@ -236,7 +236,7 @@
 
             $.ajax({
                 type: "POST",
-                url: "/usr/member/certification",
+                url: "/api/certification",
                 data: {"userEmail": userEmail, "inputCode": inputCode.value},
                 beforeSend: function (xhr) {
                     xhr.setRequestHeader(header, token);
